requirements.txt
amqp==2.3.2
billiard==3.5.0.5
captcha==0.3
celery==4.2.1
certifi==2018.11.29
chardet==3.0.4
cryptowatch-client==0.1.4
Django==2.1.7
django-braces==1.13.0
django-celery-results==1.0.4
django-countries==5.3.3
django-model-utils==3.1.2
django-ranged-response==0.2.0
django-simple-captcha==0.5.10
django-star-ratings==0.7.0
django-widget-tweaks==1.4.3
idna==2.8
kombu==4.2.2.post1
Pillow==5.4.1
pkg-resources==0.0.0
python-bitcoinlib==0.10.1
python-bitcoinrpc==1.0
pytz==2018.9
requests==2.21.0
six==1.12.0
swapper==1.1.0
urllib3==1.24.1
vine==1.2.0
